Tentatively remove MongoDB access from crawler
The Crawler used to use MongoDB to access a bunch of stuff from the
database when preparing its own insert.  I have tentatively changed
this to use passed parameters (which would allow us to keep the MongoDB
on localhost only on the master), but I have yet to check if it still
shows up properly in the frontend after all these changes.