Fix logic bug in OpenWPM code - ref #31
The "url" parameter was being overwritten by later SQL queries,
leading to nonsensical queries even later. This would have worked
with java-style variable scoping, but in python, this led to
problems.

Additionally, the SQLite cursor was being re-used. Since the cursor
is actually keeping state internally, this is a bad idea. I have
created an explicit outer cursor to keep state for the outermost
SQLite loop.

I have done some verification that all other results are still valid,
but it would be great if someone could check it out in more depth.
This also resolves the https bug from #31 - I was unable to reproduce
the crash mentioned there.