testssl: Add HSTS preloading checks
Check for the presence of the domain in the HSTS preloading list.
Also walks up the domain tree and looks if superdomains (is that
even a word?) are preloaded with the include-subdomains directive.

Note: HSTS preloading requires you to offer the relevant headers on
your second-level domain, but not on subdomains. So, checks on
subdomains may return no headers, even though the domain is in the
preloading list. Take that into account when designing the rating
algorithm.