ansible: Added a privacyscore__debug variable to control test environments
We only want to install certain packages when we are in a test
environment.  With this variable the dev requirements are only installed
when we are indeed in debug mode.

We currently default to True. We probably want to reconsider once the
ansible stuff is working as expected, because we should aim for being
secure by default.