install tests: make it debian stale compatible
We face the issue of pip not being able to install "any working version"
if installing one fails. Fair enough. plyvel has just released 1.0.0 and
breaks compatibility with Debian stale. Because we use that in
production, we want to stay compatible with it. For now, at least.
We thus manually install an older version of plyvel in openwpm's
virtualenv.