Let scanfromfile read sleep time from file
Having a constant sleep time is not optimal in some situations.
This commit introduces a new parameter, -f / --sleep-from-file,
that allows users to specify a file from which the time to sleep
after each scan should be read. The file should contain a single
line with a floating point number that will be passed to the
python sleep command (i.e. specified in seconds, not milliseconds).