scanfromfile: Remove celery inspection code
We cannot make use of that code, because it cannot perform what we
actually want.  As far as I have understood the docs, we cannot find out
the number of open tasks. reserved() returns the tasks that the local
node has claimed and (pre) fetched rather than what needs to processed
on a global scale.