ansible: Create separate definition for handlers
Now we don't need to systemd reload all the time, but can do it once all
the services are installed. The order of the handlers being written to
the file is important, i.e. the handler defined first is run first.
Rather than the handler that is notified first is run first.