ansible: Only refresh packages every 30 minutes
It seems that certain mirrors don't like being hit often. We try to
reduce the load by refreshing the package list after 30 minutes. That
should be sufficient for the mirrors to not be hit in subsequent runs
and for us to have a small enough timeout to start over in case stuff
broke.