Fix third party include false positives
For some reason, als URLs starting with http:// or https:// were
declared trackers.  This commit strips the protocol from the URLs.
We'll have to see if this was the only bug causing this.  The root
cause seems to lie with adblockparser, which we should replace with
the Princeton code anyway.